+++ src/pages/Mypage.jsx
@@ -1,18 +1,76 @@
-import { useState } from "react";
+import { useEffect, useState } from "react";
 import styled from "styled-components";
 import MypageButton from "../components/mypage/MypageButton";
 import MypageProfile from "../components/mypage/MypageProfile";
-import MypageContents from "../components/mypage/MypageContents";
+import MypageContentsBookmark from "../components/mypage/MypageContentsBookmark";
+import MypageContentsReview from "../components/mypage/MypageContentsReview";
+import {
+  collection,
+  onSnapshot,
+  orderBy,
+  query,
+  where,
+} from "firebase/firestore";
+import { dbService } from "../firebase";
 
 const Mypage = () => {
   const [category, setCategory] = useState("bookmark"); // review, bookmark
+  const [bookmark, setBookmark] = useState([]);
+  const [reviews, setReviews] = useState([]);
+  const user = JSON.parse(localStorage.getItem("User"));
+  const bookmarkCount = bookmark.length;
+  const reviewsCount = reviews.length;
+  useEffect(() => {
+    onSnapshot(
+      query(
+        collection(dbService, "bookmark"),
+        orderBy("createdAt", "desc"),
+        where("userId", "==", user.uid)
+      ),
+      (snapshot) => {
+        const newBookmark = snapshot.docs.map((doc) => ({
+          id: doc.id,
+          ...doc.data(),
+        }));
+        setBookmark(newBookmark);
+      }
+    );
+
+    onSnapshot(
+      query(
+        collection(dbService, "reviews"),
+        orderBy("createdAt", "desc"),
+        where("userId", "==", user.uid)
+      ),
+      (snapshot) => {
+        const newReview = snapshot.docs.map((doc) => ({
+          id: doc.id,
+          ...doc.data(),
+        }));
+        setReviews(newReview);
+      }
+    );
+  }, []);
 
   return (
     <StyledDivContainer>
       <StyledDivMain>
-        <MypageProfile />
+        <MypageProfile
+          bookmarkCount={bookmarkCount}
+          reviewsCount={reviewsCount}
+        />
         <MypageButton category={category} setCategory={setCategory} />
-        <MypageContents category={category} />
+        <Contents>
+          {category === "bookmark" ? (
+            <MypageContentsBookmark category={category} bookmark={bookmark} />
+          ) : (
+            <MypageContentsReview
+              category={category}
+              reviews={reviews}
+              user={user}
+            />
+          )}
+        </Contents>
       </StyledDivMain>
     </StyledDivContainer>
   );
@@ -34,6 +92,12 @@ const StyledDivMain = styled.div`
   background: #fef6ec;
   border-radius: 30px;
   padding: 50px;
+`;
+const Contents = styled.div`
+  padding: 20px;
+  border-radius: 30px;
+  background-color: #f5f5f5;
+  flex-direction: column;
 `;
 
 export default Mypage;
